requirements.txt
Django==3.0
djangorestframework_simplejwt==4.4.0
djangorestframework==3.11.0
secrets==1.0.2
